fixup! Introduce the MouseDirection type and the applyRendererMoveChanges function on the timeline cell mouse handler and renderer
